# docs/api/openapi.yaml
components:
  schemas:
    AccountBalanceOut:
      properties:
        account_id:
          title: Account Id
          type: string
        current_balance_cents:
          title: Current Balance Cents
          type: integer
        projected_balance_cents:
          title: Projected Balance Cents
          type: integer
      required:
      - account_id
      - current_balance_cents
      - projected_balance_cents
      title: AccountBalanceOut
      type: object
    AccountCreate:
      additionalProperties: false
      properties:
        currency:
          maxLength: 3
          minLength: 3
          title: Currency
          type: string
        initial_balance_cents:
          default: 0
          minimum: 0.0
          title: Initial Balance Cents
          type: integer
        kind:
          title: Kind
          type: string
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
      required:
      - name
      - kind
      - currency
      title: AccountCreate
      type: object
    AccountOut:
      properties:
        archived_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Archived At
        created_at:
          format: date-time
          title: Created At
          type: string
        currency:
          title: Currency
          type: string
        current_balance_cents:
          title: Current Balance Cents
          type: integer
        current_balance_version:
          title: Current Balance Version
          type: integer
        id:
          format: uuid
          title: Id
          type: string
        initial_balance_cents:
          title: Initial Balance Cents
          type: integer
        kind:
          title: Kind
          type: string
        name:
          title: Name
          type: string
      required:
      - id
      - name
      - kind
      - currency
      - initial_balance_cents
      - current_balance_cents
      - current_balance_version
      - archived_at
      - created_at
      title: AccountOut
      type: object
...
    Body_login_auth_login_post:
      properties:
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
        grant_type:
          anyOf:
          - pattern: ^password$
            type: string
          - type: 'null'
          title: Grant Type
        password:
          format: password
          title: Password
          type: string
        scope:
          default: ''
          title: Scope
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - password
      title: Body_login_auth_login_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RegisterRequest:
      additionalProperties: false
      properties:
        email:
          format: email
          title: Email
          type: string
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        password:
          maxLength: 128
          minLength: 1
          title: Password
          type: string
      required:
      - email
      - password
      - name
      title: RegisterRequest
      type: object
    ReversalCreate:
      additionalProperties: false
      properties:
        expected_version:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Expected Version
        idempotency_key:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
      required:
      - idempotency_key
      title: ReversalCreate
      type: object
...
    TokenPair:
      additionalProperties: false
      properties:
        access_token:
          title: Access Token
          type: string
        refresh_token:
          title: Refresh Token
          type: string
        token_type:
          default: bearer
          title: Token Type
          type: string
      required:
      - access_token
      - refresh_token
      title: TokenPair
      type: object
    TransactionCreate:
      additionalProperties: false
      properties:
        amount_cents:
          exclusiveMinimum: 0.0
          title: Amount Cents
          type: integer
        description:
          anyOf:
          - maxLength: 500
            minLength: 1
            type: string
          - type: 'null'
          title: Description
        external_id:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: External Id
        fingerprint:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Fingerprint
        idempotency_key:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
        occurred_at:
          format: date-time
          title: Occurred At
          type: string
        operation_type:
          title: Operation Type
          type: string
      required:
      - idempotency_key
      - operation_type
      - amount_cents
      - occurred_at
      title: TransactionCreate
      type: object
    TransactionOut:
      properties:
        amount_cents:
          title: Amount Cents
          type: integer
        balance_after_cents:
          default: 0
          title: Balance After Cents
          type: integer
        balance_version:
          default: 0
          title: Balance Version
          type: integer
        category_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Category Id
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Fingerprint
        id:
          format: uuid
          title: Id
          type: string
        kind:
          title: Kind
          type: string
        occurred_at:
          format: date-time
          title: Occurred At
          type: string
        operation_type:
          title: Operation Type
          type: string
        reversal_of_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Reversal Of Id
        status:
          title: Status
          type: string
      required:
      - id
      - kind
      - operation_type
      - status
      - amount_cents
      - occurred_at
      - description
      - external_id
      - fingerprint
      - reversal_of_id
      - category_id
      - created_at
      title: TransactionOut
      type: object
    TransferCreate:
      additionalProperties: false
      properties:
        amount_cents:
          exclusiveMinimum: 0.0
          title: Amount Cents
          type: integer
        idempotency_key:
          maxLength: 128
          minLength: 1
          title: Idempotency Key
          type: string
        occurred_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Occurred At
        target_account_id:
          format: uuid
          title: Target Account Id
          type: string
      required:
      - idempotency_key
      - amount_cents
      - target_account_id
      title: TransferCreate
      type: object
    TwoFactorChallengeResponse:
      additionalProperties: false
      properties:
        challenge_id:
          title: Challenge Id
          type: string
        two_factor_required:
          default: true
          title: Two Factor Required
          type: boolean
      required:
      - challenge_id
      title: TwoFactorChallengeResponse
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login
      type: oauth2
info:
  title: Cifra API
  version: 0.2.0
openapi: 3.1.0
paths:
  /accounts:
    get:
      operationId: list_accounts_accounts_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccountOut'
                title: Response List Accounts Accounts Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: List Accounts
      tags:
      - accounts
    post:
      operationId: create_account_accounts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Account
      tags:
      - accounts
...
  /accounts/{account_id}/balance:
    get:
      operationId: account_balance_accounts__account_id__balance_get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      - in: query
        name: projected
        required: false
        schema:
          default: false
          title: Projected
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBalanceOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Account Balance
      tags:
      - accounts
...
  /accounts/{account_id}/transactions:
    get:
      operationId: list_transactions_accounts__account_id__transactions_get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TransactionOut'
                title: Response List Transactions Accounts  Account Id  Transactions
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Transactions
      tags:
      - transactions
    post:
      operationId: create_transaction_accounts__account_id__transactions_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Transaction
      tags:
      - transactions
  /accounts/{account_id}/transactions/transfers:
    post:
      operationId: create_transfer_accounts__account_id__transactions_transfers_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TransactionOut'
                title: Response Create Transfer Accounts  Account Id  Transactions
                  Transfers Post
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Transfer
      tags:
      - transactions
  /accounts/{account_id}/transactions/{transaction_id}/reversal:
    post:
      operationId: reverse_transaction_accounts__account_id__transactions__transaction_id__reversal_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      - in: path
        name: transaction_id
        required: true
        schema:
          format: uuid
          title: Transaction Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReversalCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Reverse Transaction
      tags:
      - transactions
...
  /auth/login:
    post:
      operationId: login_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_auth_login_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TokenPair'
                - $ref: '#/components/schemas/TwoFactorChallengeResponse'
                title: Response Login Auth Login Post
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - auth
...
  /auth/register:
    post:
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - auth
...